requirements.txt
pygame-ce==2.5.0
flake8==7.0.0 